1- initialize launch.json file 2- create development main 3- create production main
