+++ .vscode/launch.json
@@ -1,0 +1,20 @@
+{
+    "version": "0.2.0",
+    "configurations": [
+      {
+        "name": "dev_mode", // this will appear when you press on run
+        "request": "launch",
+        "type": "dart",
+        "program": "lib/main_development.dart",
+        "args": ["--flavor", "development", "--target", "lib/main_development.dart" ]
+      },
+      {
+        "name": "prod_mode", // this will appear when you press on run
+        "request": "launch",
+        "type": "dart",
+        "program": "lib/main_production.dart",
+        "args": ["--flavor", "production", "--target", "lib/main_production.dart" ]
+      }
+    ],
+
+  }
