+++ lib/main_development.dart
@@ -1,0 +1,25 @@
+import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+
+void main() {
+  SystemChrome.setSystemUIOverlayStyle(const SystemUiOverlayStyle(
+    systemNavigationBarColor: Colors.red, // navigation bar color
+    statusBarColor: Colors.red, // status bar color
+  ));
+  runApp(const MyApp());
+}
+
+class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    // TODO: implement build
+    return const MaterialApp(
+      debugShowCheckedModeBanner: true,
+      home: Scaffold(
+        body: Text("This is development mode"),
+      ),
+    );
+  }
+}
